+++ src/constants/Extension.ts
@@ -1,0 +1,3 @@
+
+
+export const EXTENSION_NAME = "vscode-msgraph-autocomplete";