+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.0.3]
+
+- Speed improvement by leveraging the `globalState` from VSCode for caching
+- Added a command to clear the cache
+
 ## [0.0.2]
 
 - [#1](https://github.com/estruyf/vscode-msgraph-autocomplete/issues/1): Added support for `{user-id}` token and `/users/` API path